backend/requirements.txt

fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic<2
boto3==1.34.162
pdfminer.six==20240706
pypdf==4.3.1
pandas==2.2.3
numpy==2.1.2
psycopg[binary]==3.2.1
pinecone-client==5.0.1
httpx==0.27.2
sqlalchemy
psycopg2-binary
alembic
reportlab==4.0.7
